🧪 Make post-gate job chain run on gate success
Previously, jobs following the `check` job didn't have `if` conditionals
or the existing expressions didn't use the status check functions. Those
jobs are responsible for the finishing steps of the release automation.
They used to work when the entire parent job tree was successful. The
release automation is triggered by the `workflow_dispatch` event, which
has a boolean flag/checkbox for disregarding the testing stage status.
Using it makes the `check` gate pass even when some test matrix jobs
fail. It was not obvious that jobs the depend on `check` don't run if
any of the parent chain jobs fail. Not by default, so they were being
skipped for no obvious reason.

This patch updates said `if` conditionals to force GitHub Actions CI/CD
to only take into account the `check` job result and not its
dependencies by prepending the trick of
`always() && needs.check.result == 'success' &&` in front of each `if`
conditional in publishing jobs.

Refs:
* https://github.com/actions/runner/issues/2205#issuecomment-1534430719
* https://github.com/orgs/community/discussions/45058#discussioncomment-7465690
* https://docs.github.com/en/actions/learn-github-actions/expressions#status-check-functions

Co-Authored-By: Pavel Březina <pbrezina@redhat.com>